changed type from short to int of attributes

+++ src/org/dinopolis/gpstool/gpsinput/garmin/GarminWaypoint.java
@@ -100,7 +100,7 @@ public interface GarminWaypoint
  * @throws UnsupportedOperationException
  */
 	
-	public short getAttributes()
+	public int getAttributes()
     throws UnsupportedOperationException;
